requirements.txt
blinker==1.6.2
click==8.1.3
Flask==2.3.2
Flask-Cors==4.0.0
importlib-metadata==6.7.0
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.3
mysql-connector-python==8.0.33
protobuf==3.20.3
python-dotenv==1.0.0
Werkzeug==2.3.6
zipp==3.15.0
